app/requirements.txt
Flask==3.0.2
keras==2.15.0
numpy==1.24.3
# Pillow==9.4.0
Pillow==10.3.0
Gunicorn==21.2.0
Tensorflow
Streamlit